Todo exercise
Added functionality: (script.js) defined variables, event listeners, and built a function that creates a new todo using a Post request and URLSearchParams. Same for edit.js but I have two functions one to display data based on a todo id and the other allows you to update that information.